requirements.txt
requests
sqlalchemy
psycopg2-binary
python-dotenv
schedule
streamlit
plotly
pandas
